requirements.txt
ffmpeg-python
perceptron
pydantic
python-dotenv
